Move rest of river/internal to river
